% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAGen.R
\name{DNA}
\alias{DNA}
\title{DNA Generator}
\usage{
DNA(n, control, t)
}
\arguments{
\item{n, }{control, t}
}
\value{
DNAStringGen[i]
}
\description{
Generates randomized DNA sequences
}
\examples{
DNA(4,control=1,2)

}
