% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairwiseAli.R
\name{gen_matrix}
\alias{gen_matrix}
\title{Pairwise Alignment}
\usage{
gen_matrix(str1, str2, m, s, g)
}
\arguments{
\item{str1, str2, m, s, g}{}
}
\value{
numbr_of_sim
}
\description{
Pairwise alignment function
}
\examples{
gen_matrix(str1,str2,m=2,s=2,g=3)

}
