% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hamming_distance}
\alias{hamming_distance}
\title{HDDS}
\usage{
hamming_distance(s1, s2, gapO, gapE)
}
\arguments{
\item{s1, s2, gapO, gapE}{}
}
\value{
Total_score
}
\description{
This package is meant for play-test and interactive presentation.
}
\examples{
hamming_distance(s1,s2,gapO,gapE)

}
