% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Hamming Distance}
\usage{
hamming_distance(s1, s2, gapO, gapE)
}
\arguments{
\item{s1, s2, gapO, gapE}{}
}
\value{
total_score
}
\description{
Hamming Distance modified to for biology for aligning biological sequences. However, this is still a prototype.
}
\examples{
hamming_distance(s1,s2,2,2)

}
